int 32 to 64 fix #2

+++ CQMacroCreator/Form1.cs
@@ -1632,7 +1632,7 @@ namespace CQMacroCreator
             }
             else
             {
-                upperCount.Value = (int)(d * Int32.Parse(followerLabel.Text.Replace(" ", "")));
+                upperCount.Value = (long)(d * Int64.Parse(followerLabel.Text.Replace(" ", "")));
             }
         }
 
@@ -1644,7 +1644,7 @@ namespace CQMacroCreator
             }
             else
             {
-                lowerCount.Value = (int)(d * Int32.Parse(followerLabel.Text.Replace(" ", "")));
+                lowerCount.Value = (long)(d * Int64.Parse(followerLabel.Text.Replace(" ", "")));
             }
         }
 
@@ -1657,7 +1657,7 @@ namespace CQMacroCreator
             }
             else
             {
-                lowerCount.Value = (int)(0.3 * Int32.Parse(followerLabel.Text.Replace(" ", "")));
+                lowerCount.Value = (long)(0.3 * Int64.Parse(followerLabel.Text.Replace(" ", "")));
             }
         }
